+++ src/main/java/com/besson/arknights/block/ModBlockEntities.java
@@ -3,6 +3,7 @@ package com.besson.arknights.block;
 import com.besson.arknights.ArknightsFurniture;
 import com.besson.arknights.block.columbian.RedBrownCabinetBE;
 import com.besson.arknights.block.fantastic.SmallSquareNightstandBE;
+import com.besson.arknights.block.lgd.SmallSandalwoodNightstandBE;
 import com.besson.arknights.block.sideline.CheckerboardBookcaseBE;
 import com.besson.arknights.block.sideline.NightstandBE;
 import com.besson.arknights.block.sideline.SimpleCabinetBE;
@@ -73,6 +74,10 @@ public class ModBlockEntities {
     public static final RegistryObject<BlockEntityType<SmallSquareNightstandBE>> SMALL_SQUARE_NIGHTSTAND =
             register("small_square_nightstand", SmallSquareNightstandBE::new,
                     () -> new Block[]{ModBlocks.SMALL_SQUARE_NIGHTSTAND.get()});
+
+    public static final RegistryObject<BlockEntityType<SmallSandalwoodNightstandBE>> SMALL_SANDALWOOD_NIGHTSTAND =
+            register("small_sandalwood_nightstand", SmallSandalwoodNightstandBE::new,
+                    () -> new Block[]{ModBlocks.SMALL_SANDALWOOD_NIGHTSTAND.get()});
 
     private static <T extends Entity> RegistryObject<EntityType<T>> register(String name, EntityType.Builder<T> builder)
     {